shared lib: don't export unnecessary structures to user
internal.h is dedicated to hold the non-user functions and structures for shared
lib.
